repr: move Timestamp, Diff, and supporting impls out of root
Having these types in the crate root makes the crate a bit messy, as you
can't easily tell what's reexported publicly and what's just imported to
support the impls on the Timestamp type.

Move both of Timestamp and Diff into their own modules, to make the
structure clearer.
